rx_deframer: Pipeline received data byte by 1 clock
We need to have the data available at our outputs when byte_available
is asserted, which happens in the cycle after we receive the 8th bit.
It is awkward to move byte_available a cycle earlier because we don't
know until we see the 8th bit whether we are actually dealing with
a flag.  So, add an extra latch to make the data line up with the
strobe.